Remove default brew installed conflicts and separate the build process.
Removes jpeg-turbo and freetype from the brew install, using -f to force, so that they can't accidently be used by our compile process.

Breaks the 2_build.sh --build all step into multiple steps, so that then each one becomes a step in the build process, and it becomes easier to see the logs of what is happening, if/when there's issues.